[mod] Manage UI Revamp
create, edit, test, settings
